verc3/synthesis/enumerate: Clarify and fix thread-safety
Previously RangeEnumerate::Extend would return kInvalidID if called
again. This was the case if 2 threads concurrently wanted to call
LambdaOptions::Register, but since RangeEnumerate::Extend only succeeds
for one, one would not update the current id. Note that, here it was
also possible that the thread that obtained the id in
LambdaOptions::Register may not have completed the write of id_, but the
thread that failed would then continue and may be possible to still read
an invalid id_. This was a reasonable restriction, given that
RangeEnumerate::operator[] should not be used concurrently with Extend,
but was unnecessarily confusing (one uncommon case exists where this is
still valid, i.e. if a RangeEnumerate instance is exclusive to a
LambdaOption).

Since this change no longer returns kInvalidID upon repeated extension
(with same label), and instead throws an exception (to make sure that an
extension does not happen accidentally with the same label), the
LambdaOption::Register implementation was changed to double-checked
locking.

Furthermore, this clarifies the thread-safety situation.