Revise interface for dealing with error traces
Revise interface to be able to obtain an ErrorTrace at a later point
(via a ErrorHashTrace), e.g. in case not all evaluated models' error
traces are useful. Obtaining just the trace of state hashes
(ErrorHashTrace) is comparatively cheap.

In addition, to be able to store all information to provide complete
debugging information, Properties may print errors only after the model
has been successfully evaluated (no ErrorHashTrace); a Clone method was
added to be able to store Properties of interest.