verc3/core: Add support for Unknowns
Unknowns can be used in place where the model encounters a situation
where the implementation is not fully specified and no further progress
can be made. This allows checking incomplete specifications as far as
possible.

Additionally, Properties are made aware of Unknowns encountered, such
that they are able to avoid false positives.